requirements_windows.txt
# Sign-KID Windows环境依赖包
# 基于原requirements.txt，添加缺失包，移除Windows不兼容包

# ============================================
# 核心深度学习框架
# ============================================
torch==1.3.0
torchtext==0.4.0
tensorboard==1.14.0

# ============================================
# 数据处理和科学计算
# ============================================
numpy==1.17.3
scipy==1.3.1
pandas==0.25.1

# ============================================
# 配置和工具
# ============================================
pyyaml==5.4.1
tqdm==4.36.1

# ============================================
# 新增：核心依赖（原requirements.txt缺失）
# ============================================
einops>=0.3.0  # diffusion.py中使用
opencv-python>=4.5.0  # plot_videos.py中使用（cv2）

# ============================================
# GUI相关（如果需要图形界面）
# ============================================
PyQt5>=5.15.0  # GUI界面
psycopg2-binary>=2.9.0  # PostgreSQL数据库（main.py中使用）

# ============================================
# 可选依赖（用于dtw.py测试代码）
# ============================================
scikit-learn>=0.24.0  # dtw.py中可选使用
nltk>=3.6  # dtw.py中可选使用

# ============================================
# 其他工具包（从原requirements.txt保留）
# ============================================
future==0.18.1
six==1.12.0
pillow==6.2.0
matplotlib==3.1.1

# ============================================
# Windows不兼容的包（已注释，不要安装）
# ============================================
# pyinotify==0.9.6  # Linux专用，Windows不支持

# ============================================
# 调试工具（可选）
# ============================================
# ipdb  # 调试工具，可选安装
# pip install ipdb

